added user owner guard

+++ ChatTutor/core/blueprints/bp_data/data.py
@@ -167,12 +167,15 @@ def delete_doc():
     data = request.json
     collection_name = data["collection"]
     doc_name = data["doc"]
-    # TODO: check if collection is owned by the user
-    collection = db.client.get_collection(name=collection_name)
-    print(collection)
-    collection.delete(where={"from_doc": doc_name})
-    print("deleted")
-    return jsonify({"deleted": doc_name, "from_collection": collection_name})
+    if DataBase().validate_course_owner(
+        collectionname=collection_name, user_email=flask_login.current_user.email
+    ):
+        collection = db.client.get_collection(name=collection_name)
+        print(collection)
+        collection.delete(where={"from_doc": doc_name})
+        print("deleted")
+        return jsonify({"deleted": doc_name, "from_collection": collection_name})
+    return jsonify({"error": "Unauthorized operation."})
 
 
 @data_bp.route("/add_doc_tosection", methods=["POST"])
@@ -207,12 +210,16 @@ def add_fromdoc_tosection():
     collection_name = data["collection"]
     section_id = data["section_id"]
     url_to_add = data["url_to_add"]
-    # TODO: check if collection is owned by the user
-    DataBase().update_section_add_fromdoc(section_id=section_id, from_doc=url_to_add)
-    return jsonify({"added": url_to_add, "to_collection": collection_name})
+    if DataBase().validate_course_owner(
+        collectionname=collection_name, user_email=flask_login.current_user.email
+    ):
+        DataBase().update_section_add_fromdoc(section_id=section_id, from_doc=url_to_add)
+        return jsonify({"added": url_to_add, "to_collection": collection_name})
+    return jsonify({"error": "Unauthorized operation."})
 
 
 @data_bp.route("/get_section", methods=["POST"])
+@flask_login.login_required
 def get_section():
     """
     The function `get_section` retrieves a specific section from a collection and returns the section
@@ -235,10 +242,13 @@ def get_section():
         }
         ```
     """
-    # TODO: check if collection is owned by the user
     data = request.json
     collection_name = data["collection"]
     section_id = data["section_id"]
-    sections, _ = DataBase().get_sections_by_id(section_id=section_id)
-    pfrom = [s.pulling_from for s in sections]
-    return jsonify({"sections": sections, "pulling_from": pfrom})
+    if DataBase().validate_course_owner(
+        collectionname=collection_name, user_email=flask_login.current_user.email
+    ):
+        sections, _ = DataBase().get_sections_by_id(section_id=section_id)
+        pfrom = [s.pulling_from for s in sections]
+        return jsonify({"sections": sections, "pulling_from": pfrom})
+    return jsonify({"error", "Unauthorized operation"})
